# backend/internal/api/docs/swagger.yaml
basePath: /api/v1
...
info:
  contact: {}
  description: |-
    Platform API to order and manage cloud resources — managed databases, apps, VMs (app servers), AI models, and public API endpoints — on the DADA Cloud platform.
    Mutations are asynchronous: create/update/delete enqueue an operation and return 202 Accepted with an operation object; poll GET /projects/{projectId}/operations/{operationId} until the operation reaches a terminal status (Ready or Failed).
  title: DADA Cloud Console API
  version: "1.0"
paths:
...
  /projects/{projectId}/builds/{buildId}:
    get:
      description: Returns a single build by id. Read-only.
      operationId: getBuild
      parameters:
      - description: Project UUID
        in: path
        name: projectId
        required: true
        type: string
      - description: Build UUID
        in: path
        name: buildId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: object with the build
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a build
      tags:
      - build
...
  /projects/{projectId}/builds/{buildId}/cancel:
    post:
      description: Cancels a queued or in-progress build. Returns 409 if the build
        already reached a terminal state. Requires write access.
      operationId: cancelBuild
      parameters:
      - description: Project UUID
        in: path
        name: projectId
        required: true
        type: string
      - description: Build UUID
        in: path
        name: buildId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: object with the updated build
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Cancel a build
      tags:
      - build
  /projects/{projectId}/builds/{buildId}/logs-token:
    post:
      description: Issues a short-lived delegate token plus WebSocket URL the frontend
        uses to stream a build's logs from the build-agent. Requires write access.
        Returns 503 when the build-agent is not configured.
      operationId: createBuildLogsToken
      parameters:
      - description: Project UUID
        in: path
        name: projectId
        required: true
        type: string
      - description: Build UUID
        in: path
        name: buildId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: object with token and ws_url
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Issue a token for the build log stream
      tags:
      - build
...
  /projects/{projectId}/environments/{envId}/apps/{appName}/builds:
    get:
      description: Returns the build history for an app in an environment, most recent
        first. Read-only.
      operationId: listBuilds
      parameters:
      - description: Project UUID
        in: path
        name: projectId
        required: true
        type: string
      - description: Environment UUID
        in: path
        name: envId
        required: true
        type: string
      - description: App name
        in: path
        name: appName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: object with a builds array
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List builds for an app
      tags:
      - build
    post:
      description: Queues a new build of the app's linked repository at its production
        branch HEAD. Builds are imperative — this returns 202 with the queued build
        (not an operation). Requires write access.
      operationId: triggerBuild
      parameters:
      - description: Project UUID
        in: path
        name: projectId
        required: true
        type: string
      - description: Environment UUID
        in: path
        name: envId
        required: true
        type: string
      - description: App name
        in: path
        name: appName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: object with the queued build
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Trigger a manual build
      tags:
      - build
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"